Map Label:  - separated code into nucleus / gl_engine  - only one draw call for all labels and implemented instancing
