+++ gl_engine/CMakeLists.txt
@@ -25,10 +25,6 @@ project(alpine-renderer-gl_engine LANGUAGES CXX)
 option(ALP_ENABLE_SHADER_NETWORK_HOTRELOAD "reload shader files per HTTP request. Allows Hot-Reload on Android/WebGL" OFF)
 set(ALP_SHADER_NETWORK_URL "http://localhost:5500/" CACHE STRING "url location, which points to the shaders directory for network hot reload")
 
-
-alp_add_git_repository(stb_slim URL https://github.com/LucasDworschak/stb_slim.git COMMITISH origin/main)
-
-
 #http://localhost:5500/
 qt_add_library(gl_engine STATIC
     Framebuffer.h Framebuffer.cpp
@@ -45,11 +41,9 @@ qt_add_library(gl_engine STATIC
     ShadowMapping.h ShadowMapping.cpp
     GpuAsyncQueryTimer.h GpuAsyncQueryTimer.cpp
     MapLabelManager.h MapLabelManager.cpp
-    MapLabel.h MapLabel.cpp
-
 
 )
-target_link_libraries(gl_engine PUBLIC nucleus Qt::OpenGL stb_slim)
+target_link_libraries(gl_engine PUBLIC nucleus Qt::OpenGL)
 target_include_directories(gl_engine PRIVATE .)
 
 qt_add_resources(gl_engine "shaders"
