+++ gl_engine/shaders/labels.frag
@@ -30,6 +30,7 @@ void main() {
     }
     else
     {
-        out_Color = vec4(1.0,1.0, 1.0, 0.25);
+        out_Color = vec4(1.0,0.0, 0.0, 1.0);
     }
+
 }
