--- nucleus/map_label/AbstractMapLabelModel.h
@@ -1,32 +1,0 @@
-/*****************************************************************************
- * Alpine Terrain Renderer
- * Copyright (C) 2023 Adam Celarek
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *****************************************************************************/
-
-#pragma once
-
-#include <vector>
-
-#include "MapLabel.h"
-
-namespace nucleus::map_label {
-
-class AbstractMapLabelModel {
-public:
-    [[nodiscard]] virtual std::vector<MapLabel> data() const = 0;
-};
-
-}
